+++ tests/testthat/test-rectangle.R
@@ -41,12 +41,29 @@ test_that("doesn't simplify lists of lists", {
   expect_equal(out$a, list(list(1), list(2)))
 })
 
+test_that("can hoist out scalars", {
+  df <- tibble(
+    x = 1:2,
+    y = list(
+      list(mod = lm(mpg ~ wt, data = mtcars)),
+      list(mod = lm(mpg ~ wt, data = mtcars))
+    )
+  )
+  out <- hoist(df, y, "mod")
+  expect_equal(out$mod, list(df$y[[1]]$mod, df$y[[2]]$mod))
+})
 
 test_that("input validation catches problems", {
   df <- tibble(x = list(list(1, b = "b")), y = 1)
 
   expect_error(df %>% hoist(y), "list-column")
-  expect_error(df %>% hoist(x, "a"), "named")
+  expect_error(df %>% hoist(x, 1), "named")
+  expect_error(df %>% hoist(x, a = "a", a = "b"), "unique")
+})
+
+test_that("string pluckers are automatically named", {
+  out <- check_pluckers("x", y = "x", z = 1)
+  expect_named(out, c("x", "y", "z"))
 })
 
 # strike ------------------------------------------------------------------