+++ NEWS.md
@@ -1,5 +1,21 @@
 # tidyr (development version)
 
+* `pack()`'s first argument is now `.data` instead of `data` (#759)
+
+* Arguments that use tidy selection syntax are now clearly documented and 
+  have been updated to use tidyselect 1.0.0 (#872).
+
+* `hoist()` errors if you use duplicated column names (@mgirlich, #834).
+
+* `hoist()` now automatically names pluckers that are a single string (#837).
+
+* `drop_na()` now preserves attributes of unclassed vectors (#905).
+
+* `pivot_wider()` gains a `names_glue` argument that allows you to construct
+  output column names with a glue specification.
+
+* `pivot_wider()` can now pivot data frame columns (#926)
+
 * `chop()` now creates list-columns of class `vctrs::list_of()`. This
   helps keep track of the type in case the chopped data frame is
   empty. This allows `unchop()` to reconstitute a data frame with